[CCUBE-1432][SR] Amending the code structure and test file structure.
